* Move include statements for glew.h and glfw.h into a header file

This is to prevent "fatal error C1189: #error: gl.h included before glew.h visual studio"

* [MESSY] Add rendering in dockable ImGui window

* Comment imGui

Added comments to the Imguiwindow files,
removed imgui comments from the core application

* Clean up RendererManager.cpp + add docs to RendererManager.h

Also: Add calls to delete texture and frame buffer in DestroySystem
Removed the render buffer object because it's currently superfluous —
we're not going to be doing depth or stencil testing for this window. Can
be added back in the future.

Ref: https://www.codingwiththomas.com/blog/rendering-an-opengl-frame
buffer-into-a-dear-imgui-window?rq=imgui

* Fixed minor warnings

* Update imGuiWindow

Updated the functions OnKeyEvent and setDockingPort

Added the capability to open and close the scene view with the f4 button and from the title bar

* Fix type conversion errors

---------

+++ Purring_Engine/src/Graphics/MeshData.cpp
@@ -137,7 +137,7 @@ namespace PE
             // Check if this mesh's VAO is currently bound
             GLint out{};
             glGetIntegerv(GL_VERTEX_ARRAY_BINDING, &out);
-            if (out == m_vertexArrayObject)
+            if (static_cast<GLuint>(out) == m_vertexArrayObject)
             {
                 // Unbind this VAO
                 glBindVertexArray(0);
